Add George framework and tests
This patch adds the main parts of the George framework. There is a still
lot of work ahead that will come in other patches. The main parts are:

 - Documentation
 - more tests that will ensure George can run in parallel
 - use fullstack tests in George

At this point, the job is added to the Experimental queue, even though
I'm confident the framework is stable, I don't see a value in running it
in the check queue yet, until we cover more scenarios.

Related-bug: #1863113

Co-Authored-By: Rodolfo Alonso Hernandez <ralonsoh@redhat.com>

Change-Id: I2b06dfc7ec586d25aa3ad2cd4be45dc32f44b5ec