WIP: ovn: Detect OVN DB schema change on reconnection
OVN introduced Chassis_Private table that is used by OVN mechanism
driver for agent API. Because this table is quite new, it may happen OVN
database is updated while neutron-server and its workers are running.

This patch creates a fanout event for SB DB Connection table. Every type
OVN SB IDL is reconnected, it checks if chassis_private table is newly
present in the schema but not in the schema helper. If needed, table
used for agent healthchecks is change in runtime from Chassis to
Chassis_Private.

Change-Id: I679945b68acf391901c8602fb1828c46cd1eec55
Closes-bug: #1901527
Signed-off-by: Jakub Libosvar <libosvar@redhat.com>