Allow shared resources between physical and tunnelled networks
Since [1] it is possible to create resource providers for tunnelled
networks and model the available bandwidth for overlay networks.

In some deployments, the admin uses the same physical network interface
for a physical network and the tunnelled networks. In that case it is
complicated to correctly split the bandwidth between the inventories
of the resource providers. This patch adds the ability to create a
resource provider with traits of both the physical network and the
tunnelled networks. E.g.:

  $ openstack resource provider trait list $brex
  +----------------------------------+
  | name                             |
  +----------------------------------+
  | CUSTOM_NETWORK_TUNNEL_PROVIDER   |
  | CUSTOM_PHYSNET_PUBLIC            |
  | ...                              |
  +----------------------------------+

With this resource provider, a request comming from a port in the
physical network or an overlay network is provided by the same
inventory.

[1]https://review.opendev.org/c/openstack/neutron/+/860639

What is missing and will be added in next patches:
* Tempest tests, that will be pushed to the corresponding repository.

Closes-Bug: #1998202
Change-Id: I801e770fd8b4d6504f5407108e1d1fd848f96c09